Partial patch from SF #452266, by Jason Petrone.
This changes Pythread_start_thread() to return the thread ID, or -1
for an error.  (It's technically an incompatible API change, but I
doubt anyone calls it.)

+++ Include/pythread.h
@@ -13,7 +13,7 @@ extern "C" {
 #endif
 
 DL_IMPORT(void) PyThread_init_thread(void);
-DL_IMPORT(int) PyThread_start_new_thread(void (*)(void *), void *);
+DL_IMPORT(long) PyThread_start_new_thread(void (*)(void *), void *);
 DL_IMPORT(void) PyThread_exit_thread(void);
 DL_IMPORT(void) PyThread__PyThread_exit_thread(void);
 DL_IMPORT(long) PyThread_get_thread_ident(void);
